+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - docs: clean up `CHANGELOG.md`
 - feat: add PHP 8.4 and Magento 2.4.8 compatibilty check with opensearch support
+- feat: enhance MySQL and Search Engine checks for `mageforge:system:check` command
 - removed: removed Github Action to watch for Changelog edits
 
 ## Latest Release